refactor: remove coffeescript from build system
make appropriate path changes to requires.
use nyc for coverage calculation.
