--- .npmignore
@@ -1,12 +1,0 @@
-.gitignore
-.npmignore
-.jshintrc
-coffeelint.json
-
-Makefile
-src
-test
-coverage
-.travis.yml
-
-ims-lti*.tgz
