+++ package.json
@@ -29,23 +29,21 @@
   "bugs": {
     "url": "http://github.com/omsmith/ims-lti/issues"
   },
-  "main": "./lib/ims-lti",
+  "main": "./src/ims-lti",
   "scripts": {
-    "prepublish": "make build"
+    "test": "nyc mocha",
+    "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "dependencies": {
     "xml2js": "~0.4.0",
-    "xmlbuilder": "~2.4.0",
-    "node-uuid": "~1.4.0"
+    "xmlbuilder": "~9.0.0",
+    "uuid": "~3.1.0"
   },
   "devDependencies": {
-    "coffee-coverage": "^0.4.4",
-    "coffee-script": "^1.8.0",
-    "coveralls": "^2.11.2",
-    "mocha": "^2.1.0",
-    "mocha-lcov-reporter": "0.0.1",
-    "mocha-multi": "^0.6.0",
-    "redis": "^0.12.1",
-    "should": "^4.6.1"
+    "coveralls": "^2.12.0",
+    "mocha": "^4.0.1",
+    "nyc": "^11.3.0",
+    "redis": "^2.8.0",
+    "should": "^13.1.3"
   }
 }
