+++ src/errors.js
@@ -5,15 +5,8 @@
  */
 class ConsumerError extends Error {
   constructor(message) {
-    {
-      // Hack: trick Babel/TypeScript into allowing this before super.
-      if (false) { super(); }
-      let thisFn = (() => { this; }).toString();
-      let thisName = thisFn.slice(thisFn.indexOf('{') + 1, thisFn.indexOf(';')).trim();
-      eval(`${thisName} = this;`);
-    }
-    this.message = message;
     super(...arguments);
+    this.message = message;
   }
 }
 class ExtensionError extends Error {
@@ -31,67 +24,32 @@ class ExtensionError extends Error {
 }
 class StoreError extends Error {
   constructor(message) {
-    {
-      // Hack: trick Babel/TypeScript into allowing this before super.
-      if (false) { super(); }
-      let thisFn = (() => { this; }).toString();
-      let thisName = thisFn.slice(thisFn.indexOf('{') + 1, thisFn.indexOf(';')).trim();
-      eval(`${thisName} = this;`);
-    }
-    this.message = message;
     super(...arguments);
+    this.message = message;
   }
 }
 class ParameterError extends Error {
   constructor(message) {
-    {
-      // Hack: trick Babel/TypeScript into allowing this before super.
-      if (false) { super(); }
-      let thisFn = (() => { this; }).toString();
-      let thisName = thisFn.slice(thisFn.indexOf('{') + 1, thisFn.indexOf(';')).trim();
-      eval(`${thisName} = this;`);
-    }
-    this.message = message;
     super(...arguments);
+    this.message = message;
   }
 }
 class SignatureError extends Error {
   constructor(message) {
-    {
-      // Hack: trick Babel/TypeScript into allowing this before super.
-      if (false) { super(); }
-      let thisFn = (() => { this; }).toString();
-      let thisName = thisFn.slice(thisFn.indexOf('{') + 1, thisFn.indexOf(';')).trim();
-      eval(`${thisName} = this;`);
-    }
-    this.message = message;
     super(...arguments);
+    this.message = message;
   }
 }
 class NonceError extends Error {
   constructor(message) {
-    {
-      // Hack: trick Babel/TypeScript into allowing this before super.
-      if (false) { super(); }
-      let thisFn = (() => { this; }).toString();
-      let thisName = thisFn.slice(thisFn.indexOf('{') + 1, thisFn.indexOf(';')).trim();
-      eval(`${thisName} = this;`);
-    }
-    this.message = message;
     super(...arguments);
+    this.message = message;
   }
 }
 class OutcomeResponseError extends Error {
   constructor(message) {
-    {
-      // Hack: trick Babel/TypeScript into allowing this before super.
-      if (false) { super(); }
-      let thisFn = (() => { this; }).toString();
-      let thisName = thisFn.slice(thisFn.indexOf('{') + 1, thisFn.indexOf(';')).trim();
-      eval(`${thisName} = this;`);
-    }
-    this.message = message;
     super(...arguments);
+    this.message = message;
   }
 }
