+++ test/MemoryNonceStore.js
@@ -3,7 +3,7 @@
  * DS102: Remove unnecessary code created because of implicit returns
  * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
  */
-const MemoryNonceStore  = require('../lib/memory-nonce-store');
+const MemoryNonceStore  = require('../src/memory-nonce-store');
 const should            = require('should');
 const shared            = require('./shared');
 
@@ -14,4 +14,3 @@ describe('MemoryNonceStore', () =>
     return new MemoryNonceStore();
   })
 );
-
